requirements.txt
langchain
langgraph
python-dotenv
chromadb
langchain-openai
langchain-anthropic
ibm-watsonx-ai
ibm-watson-machine-learning
langchain-community
langchain-ollama
langchain-ibm
jq
pypdf
pymysql
cryptography
psycopg2-binary
langchain_milvus
pymilvus
databricks-langchain
langchain-google-vertexai